:recycle: migrate tests about BackgroundModifier size
